重构代码，增强日志功能
