+++ src/AlistMediaRename/api.py
@@ -3,9 +3,15 @@ from functools import wraps
 import httpx
 import pyotp
 from typing import Callable
-from .models import ApiResponseModel
-from .utils import Message, Tools
+
+from .models import ApiResponseModel, RenameTask
 from .log import HandleException
+from .output import Output
+from .utils import Tools
+
+
+sync_client = httpx.Client()
+async_client = httpx.AsyncClient()
 
 
 # 封装接口返回信息
@@ -29,7 +35,7 @@ class ApiResponse:
                     success=True,
                     status_code=rawdata["code"],
                     error="",
-                    data=rawdata["data"],
+                    data=rawdata["data"] if rawdata["data"] else {},
                     function=func.__qualname__,
                     args=args,
                     kwargs=kwargs,
@@ -38,7 +44,7 @@ class ApiResponse:
                 success=False,
                 status_code=rawdata["code"],
                 error=rawdata["message"],
-                data=rawdata["data"],
+                data=rawdata["data"] if rawdata["data"] else {},
                 function=func.__qualname__,
                 args=args,
                 kwargs=kwargs,
@@ -53,7 +59,7 @@ class ApiResponse:
                     success=True,
                     status_code=rawdata["code"],
                     error="",
-                    data=rawdata["data"],
+                    data=rawdata["data"] if rawdata["data"] else {},
                     function=func.__qualname__,
                     args=args,
                     kwargs=kwargs,
@@ -63,15 +69,14 @@ class ApiResponse:
                     success=False,
                     status_code=rawdata["code"],
                     error=rawdata["message"],
-                    data=rawdata["data"],
+                    data=rawdata["data"] if rawdata["data"] else {},
                     function=func.__qualname__,
                     args=args,
                     kwargs=kwargs,
                 )
 
-        # 判断函数是否是异步函数
         if asyncio.iscoroutinefunction(func):
-            return async_wrapper  # type: ignore
+            return async_wrapper # type: ignore
         else:
             return sync_wrapper
 
@@ -142,14 +147,11 @@ class AlistApi:
         self.password = password
         # 使用self.totp_code.now() 生成实时 TOTP 验证码
         self.totp_code = pyotp.TOTP(totp_code)
-        self.login_success = False
         self.token = ""
         self.timeout = 10
-        self._client = httpx.Client()
-        self._async_client = httpx.AsyncClient()
 
-    # @HandleException.stop_on_error
-    @Message.output_alist_login
+    @HandleException.stop_on_error
+    @Output.output_alist_login
     @HandleException.catch_exceptions
     @ApiResponse.alist_api_response
     def login(self) -> dict:
@@ -167,21 +169,23 @@ class AlistApi:
             "Password": self.password,
             "OtpCode": self.totp_code.now(),
         }
-        r = self._client.post(url=post_url, data=post_datas, timeout=self.timeout)
+        r = sync_client.post(url=post_url, data=post_datas, timeout=self.timeout)
 
         if r.status_code != 200:
-            return {"message": "Alist 网站无法连接", "code": r.status_code, "data": {}}
+            return {"message": "Alist 网站连接失败", "code": r.status_code, "data": {}}
 
         return_data = r.json()
 
         if return_data["message"] == "success":
             self.token = return_data["data"]["token"]
 
+        # 隐藏Token信息
+        return_data["data"]["token"] = "********"
         # 返回请求结果
         return return_data
 
     @HandleException.stop_on_error
-    @Message.output_alist_file_list
+    @Output.output_alist_file_list
     @HandleException.catch_exceptions
     @ApiResponse.alist_api_response
     def file_list(
@@ -213,7 +217,7 @@ class AlistApi:
             "per_page": per_page,
             "page": page,
         }
-        r = self._client.post(
+        r = sync_client.post(
             url=post_url, headers=post_headers, params=post_params, timeout=self.timeout
         )
 
@@ -236,7 +240,7 @@ class AlistApi:
         post_url = self.url + "/api/fs/rename"
         post_headers = {"Authorization": self.token}
         post_json = {"name": name, "path": path}
-        r = self._client.post(
+        r = sync_client.post(
             url=post_url, headers=post_headers, json=post_json, timeout=self.timeout
         )
 
@@ -244,7 +248,7 @@ class AlistApi:
         return r.json()
 
     def rename_list(
-        self, rename_list: list, folder_path: str, async_mode: bool = True
+        self, rename_list: list[RenameTask], async_mode: bool = True
     ) -> list[ApiResponseModel]:
         """
         批量重命名文件.
@@ -256,12 +260,12 @@ class AlistApi:
         """
 
         if async_mode:
-            return self.rename_list_async(rename_list, folder_path)
+            return self.rename_list_async(rename_list)
         else:
-            return self.rename_list_sync(rename_list, folder_path)
+            return self.rename_list_sync(rename_list)
 
     def rename_list_async(
-        self, rename_list: list, folder_path: str
+        self, rename_list: list[RenameTask]
     ) -> list[ApiResponseModel]:
         @HandleException.catch_exceptions
         @ApiResponse.alist_api_response
@@ -279,29 +283,25 @@ class AlistApi:
             post_headers = {"Authorization": self.token}
             post_json = {"name": name, "path": path}
 
-            r = await self._async_client.post(
+            r = await async_client.post(
                 url=post_url, headers=post_headers, json=post_json, timeout=self.timeout
             )
             # 获取请求结果
             return r.json()
 
-        async def async_run(
-            rename_list: list, folder_path: str
-        ) -> list[ApiResponseModel]:
+        async def async_run(rename_list: list[RenameTask]) -> list[ApiResponseModel]:
             tasks = []
-            async with self._async_client:
+            async with async_client:
                 for file in rename_list:
-                    name = Tools.replace_illegal_char(file["target_name"])
-                    path = folder_path + file["original_name"]
+                    name = Tools.replace_illegal_char(file.target_name)
+                    path = file.folder_path + file.original_name
                     tasks.append(asyncio.ensure_future(rename_async(name, path)))  # type: ignore
 
                 return await asyncio.gather(*tasks)
 
-        return asyncio.run(async_run(rename_list, folder_path))
+        return asyncio.run(async_run(rename_list))
 
-    def rename_list_sync(
-        self, rename_list: list, folder_path: str
-    ) -> list[ApiResponseModel]:
+    def rename_list_sync(self, rename_list: list[RenameTask]) -> list[ApiResponseModel]:
         """
         批量重命名文件.
 
@@ -312,13 +312,13 @@ class AlistApi:
 
         result = []
         for file in rename_list:
-            name = Tools.replace_illegal_char(file["target_name"])
-            path = folder_path + file["original_name"]
+            name = Tools.replace_illegal_char(file.target_name)
+            path = file.folder_path + file.original_name
             result.append(self.rename(name, path))
 
         return result
 
-    @Message.output_alist_move
+    @Output.output_alist_move
     @HandleException.catch_exceptions
     @ApiResponse.alist_api_response
     def move(self, names: list, src_dir: str, dst_dir: str) -> dict:
@@ -334,13 +334,13 @@ class AlistApi:
         post_url = self.url + "/api/fs/move"
         post_headers = {"Authorization": self.token}
         post_json = {"src_dir": src_dir, "dst_dir": dst_dir, "names": names}
-        r = self._client.post(
+        r = sync_client.post(
             url=post_url, headers=post_headers, json=post_json, timeout=self.timeout
         )
         # 获取请求结果
         return r.json()
 
-    @Message.output_alist_mkdir
+    @Output.output_alist_mkdir
     @HandleException.catch_exceptions
     @ApiResponse.alist_api_response
     def mkdir(self, path: str) -> dict:
@@ -354,13 +354,13 @@ class AlistApi:
         post_url = self.url + "/api/fs/mkdir"
         post_headers = {"Authorization": self.token}
         post_json = {"path": path}
-        r = self._client.post(
+        r = sync_client.post(
             url=post_url, headers=post_headers, json=post_json, timeout=self.timeout
         )
         # 获取请求结果
         return r.json()
 
-    @Message.output_alist_remove
+    @Output.output_alist_remove
     @HandleException.catch_exceptions
     @ApiResponse.alist_api_response
     def remove(self, path: str, names: list) -> dict:
@@ -376,7 +376,7 @@ class AlistApi:
         post_url = self.url + "/api/fs/remove"
         post_headers = {"Authorization": self.token}
         post_json = {"dir": path, "names": names}
-        r = self._client.post(
+        r = sync_client.post(
             url=post_url, headers=post_headers, json=post_json, timeout=self.timeout
         )
 
@@ -397,14 +397,13 @@ class TMDBApi:
         :param key: TMDB Api Key(V3)
         """
 
-        self.api_url = "https://api.themoviedb.org/3"
+        # self.api_url = "https://api.themoviedb.org/3"
+        self.api_url = api_url
         self.api_key = api_key
         self.timeout = 10
 
-        self._client = httpx.Client()
-
     @HandleException.stop_on_error
-    @Message.output_tmdb_tv_info
+    @Output.output_tmdb_tv_info
     @HandleException.catch_exceptions
     @ApiResponse.tmdb_api_response
     def tv_info(self, tv_id: str, language: str = "zh-CN") -> tuple:
@@ -419,12 +418,12 @@ class TMDBApi:
         # 发送请求
         post_url = f"{self.api_url}/tv/{tv_id}"
         post_params = {"api_key": self.api_key, "language": language}
-        r = self._client.get(post_url, params=post_params, timeout=self.timeout)
+        r = sync_client.get(post_url, params=post_params, timeout=self.timeout)
         # 获取请求结果
         return r.json(), r.status_code
 
     @HandleException.stop_on_error
-    @Message.output_tmdb_search_tv
+    @Output.output_tmdb_search_tv
     @HandleException.catch_exceptions
     @ApiResponse.tmdb_api_response
     def search_tv(self, keyword: str, language: str = "zh-CN") -> tuple:
@@ -439,13 +438,13 @@ class TMDBApi:
         # 发送请求
         post_url = f"{self.api_url}/search/tv"
         post_params = {"api_key": self.api_key, "query": keyword, "language": language}
-        r = self._client.get(post_url, params=post_params, timeout=self.timeout)
+        r = sync_client.get(post_url, params=post_params, timeout=self.timeout)
 
         # 获取请求结果
         return r.json(), r.status_code
 
     @HandleException.stop_on_error
-    @Message.output_tmdb_tv_season_info
+    @Output.output_tmdb_tv_season_info
     @HandleException.catch_exceptions
     @ApiResponse.tmdb_api_response
     def tv_season_info(
@@ -462,13 +461,13 @@ class TMDBApi:
         # 发送请求
         post_url = f"{self.api_url}/tv/{tv_id}/season/{season_number}"
         post_params = {"api_key": self.api_key, "language": language}
-        r = self._client.get(post_url, params=post_params, timeout=self.timeout)
+        r = sync_client.get(post_url, params=post_params, timeout=self.timeout)
 
         # 获取请求结果
         return r.json(), r.status_code
 
     @HandleException.stop_on_error
-    @Message.output_tmdb_movie_info
+    @Output.output_tmdb_movie_info
     @HandleException.catch_exceptions
     @ApiResponse.tmdb_api_response
     def movie_info(self, movie_id: str, language: str = "zh-CN") -> tuple:
@@ -484,13 +483,13 @@ class TMDBApi:
         # 发送请求
         post_url = f"{self.api_url}/movie/{movie_id}"
         post_params = {"api_key": self.api_key, "language": language}
-        r = self._client.get(post_url, params=post_params, timeout=self.timeout)
+        r = sync_client.get(post_url, params=post_params, timeout=self.timeout)
 
         # 获取请求结果
         return r.json(), r.status_code
 
     @HandleException.stop_on_error
-    @Message.output_tmdb_search_movie
+    @Output.output_tmdb_search_movie
     @HandleException.catch_exceptions
     @ApiResponse.tmdb_api_response
     def search_movie(self, keyword: str, language: str = "zh-CN") -> tuple:
@@ -505,7 +504,7 @@ class TMDBApi:
         # 发送请求
         post_url = f"{self.api_url}/search/movie"
         post_params = {"api_key": self.api_key, "query": keyword, "language": language}
-        r = self._client.get(post_url, params=post_params, timeout=self.timeout)
+        r = sync_client.get(post_url, params=post_params, timeout=self.timeout)
 
         # 获取请求结果
         return r.json(), r.status_code
